More size_t

+++ include/corsika/FileIndex.h
@@ -7,9 +7,9 @@ namespace corsika
     struct FileIndex
     {
         int runNumber;
-        std::vector<unsigned long int> eventHeaders;
-        std::vector<unsigned long int> eventTrailers;
-        std::vector<unsigned long int> longBlocks;
+        std::vector<size_t> eventHeaders;
+        std::vector<size_t> eventTrailers;
+        std::vector<size_t> longBlocks;
         std::map<unsigned int, unsigned int> IDToPosition;
         std::string String();
